Static code analysis tool merged into a single project.
