+++ tools/pom.xml
@@ -17,8 +17,7 @@
 
   <modules>
     <module>archetype</module>
-    <module>static-code-analysis-config</module>
-    <module>static-code-analysis-report</module>
+    <module>static-code-analysis</module>
   </modules>
 
   <build>
