--- tools/static-code-analysis-config/.classpath
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-  <classpathentry kind="src" path="src/main/java" />
-  <classpathentry kind="src" path="src/test/java" />
-  <classpathentry kind="src" path="src/test/resources" />
-  <classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER" />
-  <classpathentry kind="con" path="org.eclipse.m2e.MAVEN2_CLASSPATH_CONTAINER" />
-  <classpathentry kind="output" path="target/classes" />
-</classpath>
