--- tools/static-code-analysis-report/README.md
@@ -1,7 +1,0 @@
-# Eclipse SmartHome Static Code Analysis Report Tool
-
-This module is used to merge the individual reports from PMD, Checkstyle and FindBugs into a single .html report.
-  
-## 3rd Party
-
-- Static Code Analysis Report Tool is based completely on source files from the https://github.com/MarkusSprunck/static-code-analysis-report that are distributed under a custom license. More information can be found in the `static-code-analysis-config` LICENSE.txt file;
