--- tools/static-code-analysis-report/pom.xml
@@ -1,76 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://maven.apache.org/POM/4.0.0"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <parent>
-    <groupId>org.eclipse.smarthome</groupId>
-    <artifactId>tools</artifactId>
-    <version>0.9.0-SNAPSHOT</version>
-  </parent>
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>org.eclipse.smarthome.tools</groupId>
-  <artifactId>static-code-analysis-report</artifactId>
-  <packaging>jar</packaging>
-
-  <name>Static Code Analysis Report Tool</name>
-  <description>Tool that aggregates reports of PMD, Checkstyle and FindBugs</description>
-
-  <properties>
-    <saxonb.version>9.0</saxonb.version>
-    <log4j.version>1.2.17</log4j.version>
-    <maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
-    <maven.assembly.version>2.6</maven.assembly.version>
-    <junit.version>4.11</junit.version>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>net.sf.saxon</groupId>
-      <artifactId>Saxon-B</artifactId>
-      <version>${saxonb.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>log4j</groupId>
-      <artifactId>log4j</artifactId>
-      <version>${log4j.version}</version>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>${junit.version}</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-assembly-plugin</artifactId>
-        <version>${maven.assembly.version}</version>
-        <configuration>
-          <descriptorRefs>
-            <descriptorRef>jar-with-dependencies</descriptorRef>
-          </descriptorRefs>
-          <archive>
-            <manifest>
-              <mainClass>org.eclipse.smarthome.buildtools.report.ScanReportUtility</mainClass>
-            </manifest>
-          </archive>
-        </configuration>
-        <executions>
-          <execution>
-            <id>make-assembly</id>
-            <phase>package</phase>
-            <goals>
-              <goal>single</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
