+++ tools/static-code-analysis/src/main/java/org/eclipse/smarthome/buildtools/rules/findbugs/CustomClassNameLengthDetector.java
