+++ tools/static-code-analysis/src/main/resources/rulesets/pmd/bundle.xml
