+++ tools/static-code-analysis/src/test/java/org/eclipse/smarthome/buildtools/rules/checkstyle/test/MethodLimitCheckTest.java
