+++ tools/static-code-analysis/src/test/java/org/eclipse/smarthome/buildtools/rules/pmd/test/WhileLoopMustUseBracesRuleTest.java
