+++ tools/static-code-analysis/src/test/resources/org/eclipse/smarthome/buildtools/rules/checkstyle/test/MANIFEST.MF
