+++ tools/static-code-analysis/src/test/resources/org/eclipse/smarthome/buildtools/rules/pmd/test/xml/WhileLoopsMustUseBracesRule.xml
